Update posts.js
There's a typo on the wiki right now. Wool Boots in Dunby are said to require merchant rating level 2, however they are available for purchase at level one. This just changes the "MerchantRating" value for Wool Boots to be 1 instead of 2. If you go to the discussion page for the commerce page on the wiki, you'll see I started a thread to discuss this.